+++ lib/punchblock/translator/asterisk/component/asterisk.rb
@@ -1,0 +1,14 @@
+module Punchblock
+  module Translator
+    class Asterisk
+      module Component
+        module Asterisk
+          extend ActiveSupport::Autoload
+
+          autoload :AGICommand
+          autoload :AMIAction
+        end
+      end
+    end
+  end
+end